moved errors and fixed prelude
