+++ src/alm/mod.rs
@@ -1,3 +1,2 @@
 pub mod enums;
 pub mod loangenerator;
-pub mod traits;
