+++ src/rates/yieldtermstructure/mod.rs
@@ -1,8 +1,6 @@
-pub mod flatforwardtermstructure;
-pub mod enums;
 pub mod discounttermstructure;
-pub mod zeroratetermstructure;
+pub mod flatforwardtermstructure;
 pub mod mixedtermstructure;
-pub mod errortermstructure;
-pub mod traits;
 pub mod tenorbasedzeroratetermstructure;
+pub mod traits;
+pub mod zeroratetermstructure;
