+++ src/rates/yieldtermstructure/traits.rs
@@ -5,8 +5,6 @@ use crate::{
     time::{date::Date, period::Period},
 };
 
-use super::errortermstructure::TermStructureConstructorError;
-
 /// # YieldTermStructureTraitClone
 /// Trait for cloning a given object.
 pub trait YieldTermStructureTraitClone {
@@ -50,6 +48,18 @@ pub enum AdvanceInTimeError {
     YieldProviderError(#[from] YieldProviderError),
     #[error("TermStructureConstructorError: {0}")]
     TermStructureConstructorError(#[from] TermStructureConstructorError),
+}
+
+#[derive(Error, Debug)]
+pub enum TermStructureConstructorError {
+    #[error("Dates and discount factors should have the same size.")]
+    DatesAndDiscountFactorsSize,
+    #[error("Dates and rates should have the same size.")]
+    DatesAndRatesSize,
+    #[error("First date needs to be reference_date")]
+    FirstDateNeedsToBeReferenceDate,
+    #[error("First discount factor needs to be 1.0")]
+    FirstDiscountFactorsNeedsToBeOne,
 }
 
 /// # YieldTermStructureTrait
